+++ server/config/configresource/config_repository.go
@@ -5,7 +5,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"os"
 
 	"github.com/kubeshop/tracetest/server/pkg/id"
 	"github.com/kubeshop/tracetest/server/resourcemanager"
@@ -15,33 +14,6 @@ var Operations = []resourcemanager.Operation{
 	resourcemanager.OperationList,
 	resourcemanager.OperationGet,
 	resourcemanager.OperationUpdate,
-}
-
-type Config struct {
-	ID   id.ID  `json:"id"`
-	Name string `json:"name"`
-
-	AnalyticsEnabled bool `json:"analyticsEnabled"`
-}
-
-func (c Config) HasID() bool {
-	return c.ID.String() != ""
-}
-
-func (c Config) GetID() id.ID {
-	return c.ID
-}
-
-func (c Config) Validate() error {
-	return nil
-}
-
-func (c Config) IsAnalyticsEnabled() bool {
-	if os.Getenv("TRACETEST_DEV") != "" {
-		return false
-	}
-
-	return c.AnalyticsEnabled
 }
 
 type option func(*Repository)
